Create s3LoggerUploader.ps1
This script finds Logger mp3 files in C:\Upload, names them based on UTC time format, and uploads them to S3.  This script is meant to be scheduled to run hourly by the host OS.  The script keeps a backup of the uploaded files for 72 hours before deleting them.  Finally, the script logs all upload activity to a text file.